added optional report selector argument example: --selector 2016q3 if selector is not specified, will match on 'Current' keyword in the <li> element

+++ pyurlscrubfed.py
@@ -8,10 +8,12 @@ import PyPDF2
 import tabula
 import pandas as pd
 from requests.packages.urllib3.exceptions import InsecureRequestWarning
-import re 
+import re
+import click
+import os
 
-requests.packages.urllib3.disable_warnings(InsecureRequestWarning)
-
+AREF_HTML = 'https://www.sec.gov/divisions/investment/13flists.htm'
+AREF_HTML_LATEST_CLASS = 'blue-chevron'
 # TABULA REF: https://github.com/chezou/tabula-py
 # PAGE AREA DEFINITION
 # OPEN PDF IN PREVIEW IN MAC, USING SQUARE SELECTION & INSPECTOR
@@ -47,37 +49,76 @@ JAVA_OPTS='-Xmx2G'
 #imported table headers
 TBL_HEADER=["CUSIP NO","-","ISSUER NAME", "ISSUER DESCRIPTION", "STATUS"]
 
-page = requests.get('https://www.sec.gov/divisions/investment/13flists.htm')
+def scrub_lis (http_url, selector):
+    # Useful section of page is stored in '<div class="article-body">'
+    # Files on the page in this class are referenced under unordered lists (html ul element)
+    # latest file is also a ul element but with formatting class 'blue-chevron' applied to it
+    ## class_elements = soup.find(class_='article-content') will return section of page with all UL's in it
+    ## uls = class_elements.find_all('ul') returns list of all UL's objects:
+    ## >>> type(lis[0])
+    ##<class 'bs4.element.Tag'>
 
-# Create a BeautifulSoup object
-soup = BeautifulSoup(page.text, 'html.parser')
+    #match on the report name matching selector
+    selector_pattern =''
 
-#pull all text from 'class-file' div (ctrl+click on element > inspect)
-file_list = soup.find(class_='blue-chevron')
-file_items = file_list.find_all('a')
+    if (selector is None):
+        #by default grab Current
+        selector_pattern=r'Current'
+        print ("Report selector not specified, looking for Current...")
+    else:
+        year,quarter=selector.split('q')
+        if (not bool(re.search('\d{4}',year)) & bool(re.search('[1-4]',quarter))):
+            print ("specified selector '{SELECTOR}' does not follow [yyyy]q[q] format".format(SELECTOR=selector))
+            exit()
+        print ("Looking for {YEAR}/{QUARTER}...".format(YEAR=year,QUARTER=quarter))
+        selector_pattern=r"{QUARTER}.. quarter {YEAR}".format(QUARTER=quarter, YEAR=year)
 
-#remove garbage at the end of 'class-file'
-#garbage = soup.find(class_='fa fa-file-pdf-o')
-#garbage.decompose()
+    print ('Scrubbing URL:\t{URL}'.format(URL=http_url))
+    # retrieve the top index page 
+    page = requests.get(http_url)
+    # Create a BeautifulSoup object
+    soup = BeautifulSoup(page.text, 'html.parser')
 
-start_page = PG_START_INDEX
-for f in file_items:
-    link = 'https://www.sec.gov' + f.get('href')
-    filename = link.rsplit('/', 1)[-1]
-    print ("PDF file link:\t", link)
-    print ("PDF filename:\t", filename)
-    print ("Grabbing file...")
-    print (link)
-    r = requests.get(link, allow_redirects=True)
-    open (filename, 'wb').write(r.content)
+    #section of html page defined under 'article-content'
+    #which contains list of all UL objects (each UL object has li elements - address and description)
+    class_elements = soup.find(class_='article-content')
+    #list of all UL objects of type <class 'bs4.element.Tag'>
+    uls = []
+    uls = class_elements.find_all('ul')
+
+    lis=[]
+    #for each group of lists on the page (each UL represents collection of li's per quarter in that year)
+    for ul in uls:
+        for li in ul.findAll('li'):
+            if (bool(re.search(selector_pattern,li.text))):
+                lis.append(li)
+
+    #href belongs to 'a' tag: lis[0].a.get('href')
+    #textual description: lis[0].text
+
+
+    # class_elements = soup.find(class_=http_class)
+    # arefs = class_elements.find_all('a')
+    # return arefs
+
+    return lis
+
+def pdf2df(filename):
+    #get number of pages in the specified file
     pdfFObj = open (filename, 'rb')
     pdfReader = PyPDF2.PdfFileReader(pdfFObj)
     num_pages = pdfReader.numPages
 
-    print ("Extracting pages:\t", start_page, '-', num_pages)
-    #tabula.convert_into(filename, filename + '.csv', output_format="csv", pages= str(start_page) + '-' + str(num_pages),area=PG_AREA,columns=PG_COLS,guess=False,java_options=JAVA_OPTS)
-    df = tabula.read_pdf(filename,pages= str(start_page)+'-'+str(num_pages),area=PG_AREA,columns=PG_COLS,guess=False,java_options=JAVA_OPTS,pandas_options={'error_bad_lines':False, 'names':TBL_HEADER})
-    
+    print ("{FILENAME}: extracting pages:{PG_START}-{PG_END}".format(FILENAME=filename,PG_START=PG_START_INDEX, PG_END=num_pages))
+
+    df = tabula.read_pdf(filename,
+        pages= str(PG_START_INDEX)+'-'+str(num_pages),
+        area=PG_AREA,
+        columns=PG_COLS,
+        guess=False,
+        java_options=JAVA_OPTS,
+        pandas_options={'error_bad_lines':False, 'names':TBL_HEADER})
+
     #delete blank spaces in the first column
     df[TBL_HEADER[0]]=df[TBL_HEADER[0]].str.replace(' ','')
 
@@ -90,7 +131,67 @@ for f in file_items:
 
     extracted_row_count = len(df)
     print ("Extracted row count:\t{ROWCOUNT}".format(ROWCOUNT=extracted_row_count))
+    return df
 
-    df.to_excel(filename+'.xlsx',index=False,header=True)
-    success = True
 
+@click.command()
+@click.option(
+    '--file', '-f',
+    help='specify local pdf securities file location')
+@click.option(
+    '--selector', '-s',
+    help='specify year and quarter of the report, ie 2018q4')
+def main(file,selector):
+    """
+    USE AT YOUR OWN RISK.
+    Utility to pull down the list of Section 13(f) securities from
+    U.S. Securities and Exchange commission public website.
+    List of securities is published in PDF file format on a quarterly basis.
+    This utility automates retrieval of the Current List, and converts extracted 
+    data in XLSX format
+    """
+
+    #disable a pesky warning
+    requests.packages.urllib3.disable_warnings(InsecureRequestWarning)
+
+    # by default assume that we are going to download a fresh file
+    local_file = False
+
+    #LOCAL FILE
+    if (file is not None):
+        #optional file argument was specified
+        #check if specified file exists
+        if (not os.path.isfile(file)):
+            print ("Specified file {FILE} does not exist. Exiting...".format(FILE=file))
+            quit()
+        else:
+            #local file exists, lets use it
+            filename = file
+            df=pdf2df (filename)
+            df.to_excel(filename+'.xlsx',index=False,header=True)
+
+    #REMOTE FILE
+    else:
+        #these come in relative format, ie /divisions/investment/13f/13flist2018q4.pdf
+        lis = scrub_lis(AREF_HTML, selector)
+        for li in lis:
+            # get the top level host fqdn address, as all the li elements contain 
+            # indirect sub-links to the tld
+            p = '(?P<host>http.*://[^:/ ]+).?(?P<port>[0-9]*).*'
+            host = re.search(p,AREF_HTML).group('host')
+            link = host + li.a.get('href')
+
+            filename = link.rsplit('/', 1)[-1]
+            print ("---\n{RNAME}: \t{FNAME}".format(RNAME=li.text, FNAME=filename))
+            #print ("Processing file...")
+            #print ("PDF file link:\t", link)
+            #print ("PDF filename:\t", filename)
+            
+            print (link)
+            r = requests.get(link, allow_redirects=True)
+            open (filename, 'wb').write(r.content)
+            df=pdf2df (filename)
+            df.to_excel(filename+'.xlsx',index=False,header=True)
+
+if __name__ == '__main__':
+    main()
